don't test commits we've already tested
This is just the reflection of me crystallizing in my head the meaning of
full_ranges and subranges (which is now renamed to untested_subranges).
Hopefully the commentary makes this clear now.

This is not just a question of performance; it's actually a functional issue
because the WorkerPool assumes the same rev doesn't get enqueued twice; if it
does this can lead to two worker threads clashing on the same
self._subprocesses entry and lead to confusing exceptions when they step on
each other's toes.