finish the stupid --no-worktrees thing
Honestly if you are going to the trouble of using this tool then your tests
must be really slow. Even for huge repos the cost of setting up the worktrees
is really not gonna be noticeable in comparison.

I dunno though maybe people have Git monorepos that are really extremely big.
I'll leave it in just in case.