Shoehorn in a cancelation mechanism
Turns out that KeyboardInterrupt is not really a good way to do this,
apparently it basically breaks the interpreter's general guarantees.

I dunno if that's the cause of my problems but it certainly isn't working.
Here's a rather ugly solution based on a SIGINT handler that forwards to a
threading.Event. This actually does nothing more than aborting WorkerPool.wait,
and we just hope that at that point everything cleans itself up.