WIP: Create worktrees lazily
This seems a bit fiddly, hopefully it's just because I'm sick.

At the moment in WorkerPool.__work we atomically pull from the queue, create
the subprocess, and insert it into _subprocesses. But now we wannt create the
worktree between pulling from the queue and inserting into _subprocceses,
otherwise worktrees won't be set up in parallel.

Need to think about if we can handle the case of a job being dequeued but not
yet interruptible. Proably the solution is to refactor the worker logic a bit,
maybe we need a Job object that can be canceled. Hopefully this doesn't mean
rewriting too much.